+++ esm/wrapper.js
@@ -1,4 +1,4 @@
-import ordrAPI from "../index.js";
+import ordrAPI from "../src/client.js";
 
 export default ordrAPI;
 export const Client = ordrAPI.Client;