+++ index.d.ts
@@ -1,2 +1,0 @@
-declare const _exports: typeof import("./src/client");
-export = _exports;