+++ src/websocket.d.ts
@@ -1,7 +1,0 @@
-export class WebSocket {
-    constructor(gateway: string);
-
-    gateway: string;
-
-    start(): void;
-}