+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.3.0] - 2021-08-18
+
+### Added
+
+-   nobot param in renders method
+-   api devmode
+-   showHitCounter param in newRender method
+-   defaults in API Doc
+-   added error codes 22 & 23
+
+### Changed
+
+-   ts types are now only in 1 file
+
+### Removed
+
+-   index.js
+
 ## [3.2.1] - 2021-08-17
 
 ### Added